test_2/requirements.txt
fastapi
uvicorn
bcrypt
pyjwt
pytz